Keep import paths in mart.attack.
