+++ mart/attack/composer/__init__.py
@@ -1,0 +1,5 @@
+from ...utils.imports import _HAS_TORCHVISION
+from .modular import *
+
+if _HAS_TORCHVISION:
+    from .vision import *
