+++ mart/attack/initializer/__init__.py
@@ -1,0 +1,5 @@
+from ...utils.imports import _HAS_TORCHVISION
+from .base import *
+
+if _HAS_TORCHVISION:
+    from .vision import *
