+++ mart/attack/initializer/base.py
