+++ mart/attack/objective/__init__.py
@@ -1,0 +1,6 @@
+from ...utils.imports import _HAS_TORCHVISION
+from .base import *
+from .classification import *
+
+if _HAS_TORCHVISION:
+    from .object_detection import *
