+++ mart/attack/objective/classification.py
@@ -4,20 +4,13 @@
 # SPDX-License-Identifier: BSD-3-Clause
 #
 
-import abc
 from typing import Any, Callable, Dict, Union
 
 import torch
 
-__all__ = ["Objective", "Mispredict", "RandomTarget"]
+from .base import Objective
 
-
-class Objective(abc.ABC):
-    """Objectives do not need to be differentiable so we do not inherit from nn.Module."""
-
-    @abc.abstractmethod
-    def __call__(self, preds: torch.Tensor, target: torch.Tensor) -> torch.Tensor:
-        raise NotImplementedError
+__all__ = ["Mispredict", "RandomTarget"]
 
 
 class Mispredict(Objective):
